backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.2
gunicorn==21.2.0
